+++ app/controllers/query/UserController.py
@@ -12,7 +12,7 @@ from app.utils.mongo.collections import *
 from app.utils.query import expand_ids_to_objects_post, expand_ids_to_objects_user
 
 
-# Authentication
+# Authentication handler which returns user object based on jwt token
 async def retrieve_user_handler(
     token: Token, connection: ASGIConnection[Any, Any, Any, Any]
 ) -> Optional[dict]:
@@ -23,6 +23,7 @@ async def retrieve_user_handler(
     return user
 
 
+# JWT authentication object
 jwt_auth = JWTAuth[dict](
     retrieve_user_handler=retrieve_user_handler,
     token_secret="secret",
@@ -30,14 +31,17 @@ jwt_auth = JWTAuth[dict](
 )
 
 
+# User controller for read operations
 class UserController(Controller):
     path = "/users"
 
+    # Get all users
     @get(path="/", tags=["User"])
     async def get_users(self) -> list[dict]:
         users = user_collection.find_documents({})
         return expand_ids_to_objects_user(users)
 
+    # Get user by id
     @get(path="/{id:uuid}", tags=["User"])
     async def get_user(self, id: UUID) -> dict:
         user = user_collection.find_document({"_id": str(id)})
@@ -45,6 +49,7 @@ class UserController(Controller):
             raise HTTPException(detail="User does not exist", status_code=HTTP_404_NOT_FOUND)
         return expand_ids_to_objects_user([user])
 
+    # Get user posts
     @get(path="/{id:uuid}/posts", tags=["User"])
     async def get_user_posts(self, id: UUID) -> dict:
         user = user_collection.find_document({"_id": str(id)})
@@ -53,6 +58,7 @@ class UserController(Controller):
         posts = post_collection.find_documents({"created_by_id": str(id)})
         return expand_ids_to_objects_post(posts)
 
+    # Get user followers
     @get(path="/{id:uuid}/followers", tags=["User"])
     async def get_user_followers(self, id: UUID) -> dict:
         user = user_collection.find_document({"_id": str(id)})
@@ -61,6 +67,7 @@ class UserController(Controller):
         followers = user_collection.find_documents({"followers": str(id)})
         return expand_ids_to_objects_user(followers)
 
+    # Get user following
     @get(path="/{id:uuid}/following", tags=["User"])
     async def get_user_following(self, id: UUID) -> dict:
         user = user_collection.find_document({"_id": str(id)})
@@ -69,11 +76,13 @@ class UserController(Controller):
         following = user_collection.find_documents({"following": str(id)})
         return expand_ids_to_objects_user(following)
 
+    # Get logged in user followers
     @get(path="/my_followers", tags=["User"])
     async def get_my_followers(self, request: Request[dict, Token, Any]) -> dict:
         followers = user_collection.find_documents({"followers": request.auth.sub})
         return followers
 
+    # Get logged in user following
     @get(path="/my_following", tags=["User"])
     async def get_my_following(self, request: Request[dict, Token, Any]) -> dict:
         following = user_collection.find_documents({"following": request.auth.sub})