+++ app/models/user_model.py
@@ -1,16 +1,18 @@
 import json
 from typing import Annotated
 from uuid import UUID
+
 from litestar.contrib.pydantic import PydanticDTO
 from litestar.contrib.sqlalchemy.base import UUIDBase
 from litestar.contrib.sqlalchemy.dto import SQLAlchemyDTO
 from litestar.dto import DTOConfig
 from sqlalchemy import Column, ForeignKey, Table
 from sqlalchemy.orm import Mapped, relationship
+from werkzeug.security import check_password_hash, generate_password_hash
+
 from .base_for_modelling import BaseModel
 from .event_model import event_attending_associations
 from .post_model import posts_shared_association
-from werkzeug.security import generate_password_hash, check_password_hash
 
 user_followers_association = Table(
     'user_followers',
@@ -20,6 +22,7 @@ user_followers_association = Table(
 )
 
 
+# User model for command
 class User(UUIDBase):
     __tablename__ = 'users'
     username: Mapped[str]
@@ -55,11 +58,11 @@ class User(UUIDBase):
         'Event', secondary=event_attending_associations, back_populates='attending_users'
     )
 
-    def generate_hash_password(self,password):
+    def generate_hash_password(self, password):
         self.password = generate_password_hash(password)
 
-    def verify_password(self,password):
-        return check_password_hash(self.password,password)
+    def verify_password(self, password):
+        return check_password_hash(self.password, password)
 
     def to_dict_create(self):
         return {
@@ -101,16 +104,7 @@ class User(UUIDBase):
         return json.dumps(message)
 
     def __repr__(self):
-        # Do not print password?
         return f"<User: id='{self.id}', username='{self.username}' >"
-
-
-# class UserRepository(SQLAlchemyAsyncRepository[User]):
-#    """User repository"""
-#
-
-
-# class UserService(Service[User]):
 
 
 write_config = DTOConfig()  # Create a DTOConfig instance for write operations